+++ pubsub/iterator.go
@@ -246,11 +246,13 @@ func (it *messageIterator) receive(maxToPull int32) ([]*Message, error) {
 	}
 	deadline := it.ackDeadline()
 	it.mu.Unlock()
-	if len(ackIDs) > 0 {
-		if !it.sendModAck(ackIDs, deadline) {
-			return nil, it.err
+	go func() {
+		if len(ackIDs) > 0 {
+			// Don't check the return value of this since modacks are fire and forget,
+			// meaning errors should not be propagated to the client.
+			it.sendModAck(ackIDs, deadline)
 		}
-	}
+	}()
 	return msgs, nil
 }
 
@@ -467,7 +469,7 @@ func (it *messageIterator) sendModAck(m map[string]bool, deadline time.Duration)
 		// transient (since the deadline is relative to the current time) and it
 		// isn't crucial for correctness (since expired messages will just be
 		// resent).
-		cctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
+		cctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
 		defer cancel()
 		bo := gax.Backoff{
 			Initial:    100 * time.Millisecond,